requirements_dev.txt
pytest>=6.0
pytest-cov>=2.0
mypy>=0.910
flake8>=3.9
tox>=3.24
pandas
numpy
biopython
ete3
torch
pytorch-lightning
torchmetrics
tensorboard
tqdm
scikit-learn